==> 6_sentiments_analysis.ipynb <==
#### Sentiment Analysis of Copom Minutes

This notebook implements a sentiment analysis procedure on the lemmatized text data from the Copom minutes.
The Dictionary used is the Loughran and McDonald.

The key steps performed in this notebook are:
 * It loads the Loughran-McDonald dictionary, which includes lists of words classified by sentiment. The words are then lemmatized to match the pre-processed text from the `lda_results.xlsx` file.
 * The script iterates through each document and counts the occurrences of positive, negative, and uncertainty words. It also includes a rule-based check for negation (`no`, `not`, etc.) to invert the sentiment of a word when preceded by a negating term.
 * The sentiment scores are appended as new columns to the existing `lda_results` DataFrame.
 * The final DataFrame, containing both topic model results and sentiment scores, is saved as a new Excel file named `sentiments_results.xlsx`.

import pandas as pd
import nltk
from nltk.stem import WordNetLemmatizer

def get_wordnet_pos(treebank_tag):
    if treebank_tag.startswith('J'):
        return nltk.corpus.wordnet.ADJ
    elif treebank_tag.startswith('V'):
        return nltk.corpus.wordnet.VERB
    elif treebank_tag.startswith('N'):
        return nltk.corpus.wordnet.NOUN
    elif treebank_tag.startswith('R'):
        return nltk.corpus.wordnet.ADV
    else:
        return nltk.corpus.wordnet.NOUN
    
    
def lemmatize_word(word, lemmatizer):
    pos_tag = nltk.pos_tag([word])[0][1]
    wordnet_pos = get_wordnet_pos(pos_tag)
    return lemmatizer.lemmatize(word, pos=wordnet_pos)

def get_loughran_mcdonald_sets(file_path="./data/Loughran-McDonald_MasterDictionary_1993-2024.xlsx"):

    lm_df = pd.read_excel(file_path)

    lm_df.columns = [col.lower() for col in lm_df.columns]
    lm_df['word'] = lm_df['word'].str.lower()

    sentiments = ["positive", "negative", "uncertainty"]
    results = []
    lemmatizer = WordNetLemmatizer()


    for sentiment in sentiments:
        words_list = lm_df[lm_df[sentiment] > 0]["word"].dropna().tolist()
        words_list_lemm = [lemmatize_word(word, lemmatizer) for word in words_list]
        results.append(set(words_list_lemm))
    
    #"positive", "negative", "uncertainty"
    return results[0], results[1], results[2]

positive_words, negative_words, uncertain_words = get_loughran_mcdonald_sets()

def analyze_text_sentiments(lemm_text, positive_set, negative_set, uncertain_set):

    negation_set = ['no', 'not', 'never', 'none', 'neither', 'nor', 'without']

    if not isinstance(lemm_text, str):
        return 0, 0, 0, 0
        
    words = nltk.word_tokenize(lemm_text.lower())
    
    total_words = len(words)
    positive_count = 0
    negative_count = 0
    
    uncertainty_count = sum(1 for word in words if word in uncertain_set)

    for i, word in enumerate(words):
        is_negated = False

        if i > 0:
            if words[i-1] in negation_set:
                is_negated = True

        if word in positive_set:
            if is_negated:
                negative_count += 1
            else:
                positive_count += 1

        elif word in negative_set:
            if is_negated:
                positive_count += 1 
            else:
                negative_count += 1
    
    return total_words, positive_count, negative_count, uncertainty_count

lda_results = pd.read_excel("./data/processed/lda_results.xlsx")

list_total_words = []
list_positive_count = []
list_negative_count = []
list_uncertainty_count = []

for index, row in lda_results.iterrows():
    lemm_text = row["lemm_text"]
    total_words, positive_count, negative_count, uncertainty_count = analyze_text_sentiments(lemm_text, positive_words, negative_words, uncertain_words)

    list_total_words.append(total_words)
    list_positive_count.append(positive_count)
    list_negative_count.append(negative_count)
    list_uncertainty_count.append(uncertainty_count)

sentiments_results = lda_results.copy()

sentiments_results['total_words'] = list_total_words
sentiments_results['positive_count'] = list_positive_count
sentiments_results['negative_count'] = list_negative_count
sentiments_results['uncertainty_count'] = list_uncertainty_count

sentiments_results.describe().round(2)

sentiments_results.to_excel('./data/processed/sentiments_results.xlsx', index=False)
