==> 10.0_regression_on_vol_nw.ipynb <==
#### Regression Analysis on Interest Rate Volatility - Newey-West.

This notebook performs a regression analysis to model how central bank communication impacts the volatility of the term structure of futures interest rates.
 * It uses several linear regression specifications to examine the relationship between different aspects of the communication—such as the event of the release, the level of the sentiment indices (optimism and uncertainty), and dummies for pessimistic communication—and the volatility of interest rates across various maturities.

import pandas as pd
import numpy as np
import statsmodels.api as sm

from sklearn.linear_model import LassoCV
from sklearn.preprocessing import StandardScaler

INITIAL_DATE = "2003-06-26"

minutes_info = pd.read_excel("./data/raw/minutes_info.xlsx")
minutes_info = minutes_info[minutes_info["DataReferencia"] >= INITIAL_DATE]
minutes_info = minutes_info.sort_values("DataReferencia", ascending=True)

dates_ref = minutes_info["DataReferencia"]
detes_pub = minutes_info["DataPublicacao"]

futures_ir = pd.read_excel("./data/future_interest_rates/future_interest_rates.xlsx", index_col="Date")
futures_ir.index = pd.to_datetime(futures_ir.index)
futures_ir = futures_ir[futures_ir.index >= INITIAL_DATE]

futures_ir.sort_index(inplace=True)

futures_ir_diff = futures_ir.diff().dropna()

vertix_names = futures_ir_diff.columns.to_list()

#### Futures MA Volatility (5 bd)

def calc_vol(returns: pd.Series) -> float:
    vol = returns.std()
    return vol

mw_volatility_5 = futures_ir_diff.rolling(5).apply(calc_vol)
mw_volatility_5 = mw_volatility_5.dropna()

mw_volatility_5 = np.log(mw_volatility_5)

mw_volatility_5["dummy_minute_pub_date"] = 0
mw_volatility_5.loc[mw_volatility_5.index.isin(detes_pub), "dummy_minute_pub_date"] = 1

#### Regressions on optimism index

df_optimism = pd.read_excel("./data/processed/index_optimism.xlsx", index_col=0)
df_optimism = df_optimism.drop(columns=['minute'])
df_optimism.columns = [f"optimism_{i}" for i in range(len(df_optimism.columns))]

df_optimism_dif = df_optimism.diff()

df_reg_opt = mw_volatility_5.copy()
df_reg_opt = df_reg_opt.join(df_optimism_dif)
cols_optimism = [col for col in df_reg_opt.columns if 'optimism_' in col]
df_reg_opt[cols_optimism] = df_reg_opt[cols_optimism].fillna(0)

len(df_reg_opt)

results_reg_opt = []
possible_y = vertix_names
all_x_reg_opt = cols_optimism + ["dummy_minute_pub_date"]
X_reg_opt_with_dummy = df_reg_opt[all_x_reg_opt]
X_reg_opt = X_reg_opt_with_dummy[cols_optimism]

scaler_reg_opt = StandardScaler()
X_reg_opt_scaled = scaler_reg_opt.fit_transform(X_reg_opt)

X_reg_opt_scaled = pd.DataFrame(X_reg_opt_scaled, index=X_reg_opt.index, columns=X_reg_opt.columns)

for maturity in possible_y:
    Y = df_reg_opt[maturity]

    lasso_cv_model = LassoCV(cv=10, random_state=100, max_iter=10000)
    lasso_cv_model.fit(X_reg_opt_scaled, Y)

    selected_coeffs_mask = lasso_cv_model.coef_ != 0
    selected_variables_names = X_reg_opt.columns[selected_coeffs_mask].tolist()

    X_selected = X_reg_opt_with_dummy[selected_variables_names + ["dummy_minute_pub_date"]]
    X_selected_sm = sm.add_constant(X_selected)

    #ols_final_model = sm.OLS(Y, X_selected_sm).fit()
    ols_final_model = sm.OLS(Y, X_selected_sm).fit(cov_type='HAC', cov_kwds={'maxlags': 10})

    all_var_names_with_const = X_selected_sm.columns.tolist()
    ols_params = pd.Series(ols_final_model.params, index=all_var_names_with_const)
    ols_std = pd.Series(ols_final_model.bse, index=all_var_names_with_const)
    ols_pvalues = pd.Series(ols_final_model.pvalues, index=all_var_names_with_const)

    results_reg_opt.append({
        'Vertix': maturity,
        'Selected_Alpha': lasso_cv_model.alpha_,
        'Coefs_OLS_Final': ols_params,
        'Std_OLS_Final': ols_std,
        'Pvalues_OLS_Final': ols_pvalues,
        'Adj_R_Sqrd_OLS_Final': ols_final_model.rsquared_adj
    })


for result in results_reg_opt:
    print(f"#### Results for Maturity (Vertex): {result['Vertix']}")
    print(f"Best alpha (λ) found by cross-validation: {result['Selected_Alpha']:.6f}")
    print(f"Adjusted R squared of the final OLS model: {result['Adj_R_Sqrd_OLS_Final']:.4f}")
            
    # --- Final OLS Model Results ---
    if not result['Coefs_OLS_Final'].empty:
        print("\n  Final OLS Results:")
        for var in result['Coefs_OLS_Final'].index:
            coef = result['Coefs_OLS_Final'][var]
            std_err = result['Std_OLS_Final'][var]
            pval = result['Pvalues_OLS_Final'][var]
            print(f"    - {var}: Coef = {coef:.6f}, Std. Err = {std_err:.6f}, P-value = {pval:.4f}")
    else:
        print("No variables selected.")
    print("-" * 60)

#### Regressions on uncertainty index

df_uncertainty = pd.read_excel("./data/processed/index_uncertainty.xlsx", index_col=0)
df_uncertainty = df_uncertainty.drop(columns=['minute'])
df_uncertainty.columns = [f"uncertainty_{i}" for i in range(len(df_uncertainty.columns))]

df_uncertainty_dif = df_uncertainty.diff()

df_reg_unc = mw_volatility_5.copy()
df_reg_unc = df_reg_unc.join(df_uncertainty_dif)
cols_unc = [col for col in df_reg_unc.columns if 'uncertainty_' in col]
df_reg_unc[cols_unc] = df_reg_unc[cols_unc].fillna(0)

results_unc = []
possible_y = vertix_names
all_x_reg_unc = cols_unc + ["dummy_minute_pub_date"]
X_reg_unc_with_dummy = df_reg_unc[all_x_reg_unc]
X_reg_unc = X_reg_unc_with_dummy[cols_unc]

scaler_reg4 = StandardScaler()
X_reg_unc_scaled = scaler_reg4.fit_transform(X_reg_unc)

X_reg_unc_scaled = pd.DataFrame(X_reg_unc_scaled, index=X_reg_unc.index, columns=X_reg_unc.columns)

for maturity in possible_y:
    Y = df_reg_unc[maturity]

    lasso_cv_model = LassoCV(cv=10, random_state=100, max_iter=10000)
    lasso_cv_model.fit(X_reg_unc_scaled, Y)

    selected_coeffs_mask = lasso_cv_model.coef_ != 0
    selected_variables_names = X_reg_unc.columns[selected_coeffs_mask].tolist()


    X_selected = X_reg_unc_with_dummy[selected_variables_names + ["dummy_minute_pub_date"]]
    X_selected_sm = sm.add_constant(X_selected)

    #ols_final_model = sm.OLS(Y, X_selected_sm).fit()
    ols_final_model = sm.OLS(Y, X_selected_sm).fit(cov_type='HAC', cov_kwds={'maxlags': 10})

    all_var_names_with_const = X_selected_sm.columns.tolist()
    ols_params = pd.Series(ols_final_model.params, index=all_var_names_with_const)
    ols_std = pd.Series(ols_final_model.bse, index=all_var_names_with_const)
    ols_pvalues = pd.Series(ols_final_model.pvalues, index=all_var_names_with_const)

    results_unc.append({
        'Vertix': maturity,
        'Selected_Alpha': lasso_cv_model.alpha_,
        'Coefs_OLS_Final': ols_params,
        'Std_OLS_Final': ols_std,
        'Pvalues_OLS_Final': ols_pvalues,
        'Adj_R_Sqrd_OLS_Final': ols_final_model.rsquared_adj
    })



for result in results_unc:
    print(f"#### Results for Maturity (Vertex): {result['Vertix']}")
    print(f"Best alpha (λ) found by cross-validation: {result['Selected_Alpha']:.6f}")
    print(f"Adjusted R squared of the final OLS model: {result['Adj_R_Sqrd_OLS_Final']:.4f}")
            
    # --- Final OLS Model Results ---
    if not result['Coefs_OLS_Final'].empty:
        print("\n  Final OLS Results:")
        for var in result['Coefs_OLS_Final'].index:
            coef = result['Coefs_OLS_Final'][var]
            std_err = result['Std_OLS_Final'][var]
            pval = result['Pvalues_OLS_Final'][var]
            print(f"    - {var}: Coef = {coef:.6f}, Std. Err = {std_err:.6f}, P-value = {pval:.4f}")
    else:
        print("No variables selected.")
    print("-" * 60)
